Code cleanup: - renamed file with space in name - removed douplicated code - refacored <any> to minimized type
