+++ src/global.d.ts
@@ -1,5 +1,1 @@
 /// <reference types="svelte" ></reference>
-
-interface Window {
-    diagramControllerInstance?: any;
-  }
