+++ src/queries/GluePointQueryBuilder.ts
